Fix printing daemon client messages
    #KT-17199 fixed

My previous commit 9819de1abd4539edb555b0c0e24f1a4090ef473f contains error.

The idea was to always print a daemon client messages when the debug log
level is enabled or when the connection was not successful.

It was the other way around in the code: when the debug log level
was disabled, all messages were printed even if connection was successful.